+++ AllHomeTasks/src/main/java/lesson8/helpers/SecondNumber.java
@@ -7,7 +7,7 @@ public class SecondNumber {
         return number2;
     }
 
-    public void setNumber2(int number2) {
+    public void setNumber2(double number2) {
         this.number2 = number2;
     }
 }
